[ZEPPELIN-3513] Zeppelin Note Creation Failure is not shown on UI
When creating a note from Zeppelin UI which fails due to permission denied exception, there is no failure/error message shown to a user on UI which is causing confusion.

This happens only in the case when HDFS is configured as the default storage for storing notebook/config.

[Bug Fix]

* [ZEPPELIN-3513](https://issues.apache.org/jira/browse/ZEPPELIN-3513)

* Does the licenses files need update? N/A
* Is there breaking changes for older versions? N/A
* Does this needs documentation? N/A

Author: Prabhjyot Singh <prabhjyotsingh@gmail.com>

Closes #3009 from prabhjyotsingh/ZEPPELIN-3513 and squashes the following commits:

db0673cc1 [Prabhjyot Singh] [ZEPPELIN-3513]change FileSystemException to IOException as HDFS throws AccessControlException when extends IOException

Change-Id: Iec5f556239160a67a55ccfba509c14ad3651c1bb